Consistently place "indirect" links after values
Old behavior meant that Direct Links set to off had the link in a
different location, but reliably after the value. With Direct Links on,
the link wrapped literals as intended, but values that are already links
got an "indirect" link, except much closer in the markup to the actual
value, and in fact actually appearing before it.

The resolution here is: whenever making an "indirect" link, it's now in
the same place: immediately after the value content itself. This means
that the links generated with Direct Links off will move somewhat from
where they were before.

There's some small accompanying styling to account for this change:
instead of trying to force itself onto a separate line, the link is
styled to just put a small margin between itself and the value.

(#52)